[#22] feat: add pre gen hook..add logging..add python version option to cookiecutter; chore: clean up gitignore
