+++ cookiecutter.json
@@ -5,6 +5,8 @@
 
     "author_email": "change.me@gmail.com",
 
+    "py_version": "3.10",
+
     "db_container_name": "db",
     "backend_container_name": "backend",
     "nginx_container_name": "nginx",
