+++ {{ cookiecutter.project_slug }}/docker-compose.yml
@@ -31,7 +31,7 @@ services:
     command: |
         bash -c "
         while !</dev/tcp/{{ cookiecutter.db_container_name }}/5432; do sleep 1; done;
-        uvicorn src.main:app --reload --workers 1 --host 0.0.0.0 --port 8000"
+        uvicorn src.main:app --reload --workers 1 --host 0.0.0.0 --port 8000 --log-config=logconfig.yml"
     depends_on:
       - {{ cookiecutter.db_container_name }}
     environment:
