+++ {{ cookiecutter.project_slug }}/{{ cookiecutter.backend_container_name }}/logconfig.yml
@@ -1,0 +1,31 @@
+version: 1
+disable_existing_loggers: False
+formatters:
+  default:
+    "()": uvicorn.logging.DefaultFormatter
+    format: '[%(asctime)s] - %(name)s - %(levelname)s - %(message)s'
+    datefmt: "%d/%b/%Y %H:%M:%S"
+  access:
+    "()": uvicorn.logging.AccessFormatter
+    format: '[%(asctime)s] - %(name)s - %(levelname)s - %(message)s'
+    datefmt: "%d/%b/%Y %H:%M:%S"
+handlers:
+  default:
+    formatter: default
+    class: logging.StreamHandler
+    stream: ext://sys.stderr
+  access:
+    formatter: access
+    class: logging.StreamHandler
+    stream: ext://sys.stdout
+loggers:
+  uvicorn.error:
+    level: INFO
+    handlers:
+      - default
+    propagate: yes
+  uvicorn.access:
+    level: INFO
+    handlers:
+      - access
+    propagate: no
