+++ {{ cookiecutter.project_slug }}/{{ cookiecutter.backend_container_name }}/pyproject.toml
@@ -5,7 +5,7 @@ description = "Base project for building fastapi backends"
 authors = ["nickatnight <nickkelly.858@gmail.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "~{{ cookiecutter.py_version }}"
 alembic = "^1.8.1"
 asyncpg = "^0.26.0"
 fastapi = "^0.81.0"
@@ -14,6 +14,7 @@ sqlmodel = "^0.0.8"
 uvicorn = "^0.18.3"
 redis = "^4.4.2"
 fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}
+PyYAML = "6.0"
 
 [tool.poetry.dev-dependencies]
 black = "22.6.0"
