+++ {{ cookiecutter.project_slug }}/{{ cookiecutter.backend_container_name }}/src/core/enums.py
@@ -1,3 +1,11 @@
+from enum import Enum
+
+
+class BaseEnum(str, Enum):
+    def __str__(self) -> str:
+        return str.__str__(self)
+
+
 class SortOrder:
     ASC = "asc"
     DESC = "desc"
