+++ {{ cookiecutter.project_slug }}/{{ cookiecutter.backend_container_name }}/src/db/session.py
@@ -1,16 +1,9 @@
-import logging
 from typing import AsyncGenerator
 
-from fastapi_cache import FastAPICache
-from fastapi_cache.backends.redis import RedisBackend
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from sqlalchemy.orm import sessionmaker
 
-from src.api.deps import get_redis_client
 from src.core.config import settings
-
-
-logger = logging.getLogger(__name__)
 
 
 engine = create_async_engine(
@@ -21,12 +14,6 @@ engine = create_async_engine(
     max_overflow=settings.MAX_OVERFLOW,
 )
 SessionLocal = sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)
-
-
-async def on_startup() -> None:
-    redis_client = await get_redis_client()
-    FastAPICache.init(RedisBackend(redis_client), prefix="fastapi-cache")
-    logger.info("FastAPI app running...")
 
 
 # async def add_postgresql_extension() -> None:
