+++ {{ cookiecutter.project_slug }}/{{ cookiecutter.backend_container_name }}/src/main.py
@@ -1,9 +1,16 @@
+import logging
+
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
+from fastapi_cache import FastAPICache
+from fastapi_cache.backends.redis import RedisBackend
 
 from src.api import routes
+from src.api.deps import get_redis_client
 from src.core.config import settings
-from src.db.session import on_startup
+
+
+logger = logging.getLogger(__name__)
 
 
 tags_metadata = [
@@ -21,11 +28,16 @@ tags_metadata = [
 
 app = FastAPI(
     title="{{ cookiecutter.project_slug }}",
-    description="base project for fastapi",
+    description="base project for fastapi backend",
     version=settings.VERSION,
     openapi_url=f"/{settings.VERSION}/openapi.json",
     openapi_tags=tags_metadata,
 )
+
+async def on_startup() -> None:
+    redis_client = await get_redis_client()
+    FastAPICache.init(RedisBackend(redis_client), prefix="fastapi-cache")
+    logger.info("FastAPI app running...")
 
 
 app.add_middleware(CORSMiddleware, allow_origins=["*"])
